requirements.txt
torch >= 1.7.1
torchvision~=0.14.1
pillow~=9.2.0
flask~=1.1.2
numpy~=1.21.5